starting to get better shape. Slow and steady.
