+++ hewo_bot_module_display/display/layouts/old_main.py
@@ -6,7 +6,8 @@ import screeninfo
 import pygame
 import os
 import sys
-
+from hewo_bot_module_display.display.face.face import Face
+from hewo_bot_module_display.settings.settings_loader import SettingsLoader
 
 class SandBox:
 
@@ -111,6 +112,16 @@ class SandBox:
         sys.exit()
 
 
-if __name__ == "__main__":
-    game = SandBox()
-    game.run()
+
+if __name__ == '__main__':
+    layout_settings = 'main'
+    display = SettingsLoader().load_settings(f'settings.layouts.{layout_settings}')
+    max_size = (display['width'], display['height'])
+    pos = [display['width'] // 5, display['height'] // 5]
+    
+    elements = [
+        Face(position=pos, enable_controls=False, max_size=max_size),
+    ]
+
+    sandbox = SandBox(elements, fullscreen=False, display=display)
+    sandbox.run()
